fyp/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
# arabic-reshaper==2.1.3
    # via xhtml2pdf
asgiref==3.3.1
    # via django
boto3==1.16.63
    # via -r requirements.in
botocore==1.19.63
    # via
    #   boto3
    #   s3transfer
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
# django-cors-headers==3.7.0
    # via -r requirements.in
django-storages==1.11.1
    # via -r requirements.in
django==3.1.6
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-storages
    #   djangorestframework
djangorestframework==3.12.2
    # via -r requirements.in
# future==0.16.0
    # via arabic-reshaper
# html5lib==1.1
    # via xhtml2pdf
idna==2.10
    # via requests
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
mysqlclient==2.0.3
    # via -r requirements.in
pillow==8.1.0
    # via
    #   -r requirements.in
    #   reportlab
    #   xhtml2pdf
# pypdf2==1.26.0
    # via xhtml2pdf
# python-bidi==0.4.2
    # via xhtml2pdf
python-dateutil==2.8.1
    # via botocore
pytz==2021.1
    # via django
# reportlab==3.5.66
    # via xhtml2pdf
requests==2.25.1
    # via -r requirements.in
s3transfer==0.3.4
    # via boto3
six==1.14.0
    # via
    #   html5lib
    #   python-bidi
    #   python-dateutil
    #   xhtml2pdf
sqlparse==0.4.1
    # via django
urllib3==1.25.11
    # via
    #   botocore
    #   requests
webencodings==0.5.1
    # via html5lib
# xhtml2pdf==0.2.5
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
